[AngryGrandpa] Add logic to give Statues of Perfection to farmhands
Added a new conditional that checks if a modded evaluation has been
done (mail flag "6324hasDoneModdedEvaluation") and the Config setting
for farmhand statues is True, it will use new logic to check for a
grandpaScore >= 4 and the lack of a "6324reward4candle" flag. If those
conditions are met it will give a Statue of Perfection reward. If the
new logic is not used, the vanilla logic is used instead.