[BugFixAddItem] Patch ItemGrabMenu.receiveLeftClick with harmony prefix
This patch experiments with invoking the menu's onAddFunction if a left
click results in Game1.player.addItemToInventoryBool returning true.
This should allow the appropriate delegate function to be called even
if players click on an item and it is auto-added to an empty slot in
inventory. Currently in the game code, delegate funnctions are not
invoked in such a case, causing bugs.