[BugFixAddItem] Apply working transpilers for right and left click.
This commit updates the transpiler patches, which are now working
properly. Each transpiler locates an insertion point in the code
where addItemToInventoryBool returned true, and an item was thus
added to inventory without invoking its delegate function. The
transpiler patch then injects a method call to OnAddItemCheck_Hook
in this mod's code - this method executes the appropriate
delegate action.