[ShowItemQuality] Use getOne() method to create dummy hud item copies
This patch corrects the issues seen in v1.1.1 by grabbing a cloned item
to use in HUD display without modifying the stack of the original item.
There is still one lingering bug where item stack overflow (e.g. already
have a stack of 100 and add 999 more) does not display the first 899
item additions before overflow in the HUD. Instead it only shows adding
the overflow quantity itself, 100, so that the updated HUD message
shows a total of 200 items (instead of 1099). However, this particular
bug also appears in the vanilla game, so I believe it's acceptable to
release this update even without fixing it yet. I may try fixing this
vanilla bug in a future release.