[AngryGrandpa] Add portrait art style config setting
This commit adds a new config property, PortraitStyle, and the
necessary components to enable it. It is included in the GMCM
settings, and all current language files have translations added
for its new options.
To accommodate the "auto" and specific portrait settings, two
instances of PortraitEditor class are now activated by the main
mod. The early one is only active for "auto" setting and selects
an appropriate portrait set to load based on other installed mods.
The other one is added late, after Content Patcher, and is only
active when PortraitStyle is not set to "auto" (present options
include "Vanilla" and "Poltergeister"). It chooses a portrait set
based on the config option selected, applying it in a way that
will overwrite any changes to grandpa's portrait that may have
been made by other mods.