[DCT] Harmony patch to space out stacked topic dialogue
This commit patches NPC.checkForNewCurrentDialogue in order to
keep track of which NPC have recently spoken to the farmer with
conversation topic dialogue. The transpiled function adds NPC
names to a RecentTopicSpeakers list when they give topic dialogue,
and conditionally skips the method's foreach block that *checks*
for topic dialogue if the NPC is already in the recent speakers
list. Other parts of the code in ModEntry clear the recent
speakers list at the start of each day, when returning to title,
and scan at every TimeChanged event to remove NPCs who were
added to the list an hour ago or more.

The current version of the transpiler patch is still fragile,
and will be made more robust with future changes.

This commit also introduces unit testing for some functions
(needs revision).